requirements.txt
streamlit
requests
beautifulsoup4
openai
pdfminer.six
python-docx
python-dotenv